refactor(web): lift SpatialEditor's largest JSX regions into sibling components
SpatialEditor.tsx was ~4057 lines with a ~1350-line inline JSX return,
making every UI change here cost a browser-test round trip instead of a
millisecond node test. Following SelectionOverlay.tsx's established
props-in/callbacks-out shape, extract the four regions that are both
large and cohesive:

- CanvasContextMenu.tsx: the right-click menu (node/edge/empty-canvas
  actions), ~570 lines of inline JSX/closures
- ConnectOverlay.tsx: the in-flight connect gesture's source indicator
  and keyboard-reachable connect targets
- MemberOutlinesOverlay.tsx: the multi-selection member/edge highlight
- HandViewControls.tsx: the dock's hand-mode zoom cluster

This is a behavior-preserving move: every expression inside each region
is textually unchanged, no new state/context/memoization is introduced,
and the conditional-render guards stay in SpatialEditor.tsx at their
original positions. SpatialEditor.tsx drops from 4057 to 3326 lines
(-731). The extracted components' props are exactly the closures/values
their region already referenced (ContextMenuTarget/LinkDialogState/
CanvasPickerState are named extractions of SpatialEditor's existing
inline useState shapes, needed to share the types across files, not new
contracts).

Resize handles were already extracted (SelectionOverlay.tsx) and are
untouched. The edge-label and group-label editors are ~30-line
TextNodeEditor wrappers, below the "substantial" bar, and are
deliberately left inline.

No new tests: the ~53 unmodified spatial-editor browser test files plus
the apps/web jsdom suite (162 files, 1952 tests) are the verification
that behavior did not change. Also verified: pnpm -r typecheck, pnpm
build, pnpm knip (clean, no orphan exports), pnpm run test:browser (97
files / 518 tests), and smoke:bundle-size (same chunk, budgets pass).

Closure-family extraction (Track 2) is a separate follow-up.
