+++ apps/web/src/components/spatial-editor/CanvasContextMenu.tsx
@@ -1,0 +1,725 @@
+/** Right-click menu: node, edge, and empty-canvas actions. */
+import type {
+  CanvasColor,
+  ClipboardFragment,
+  SpatialCanvas,
+  SpatialNode,
+} from '@kamiazya/whiteboard-canvas-model'
+import type { SpatialPresetKey } from '@kamiazya/whiteboard-canvas-render'
+import {
+  SPATIAL_DARK_PALETTE,
+  SPATIAL_LIGHT_PALETTE,
+  tidyNodes,
+} from '@kamiazya/whiteboard-canvas-render'
+import {
+  AlignCenterHorizontal,
+  AlignCenterVertical,
+  AlignEndHorizontal,
+  AlignEndVertical,
+  AlignHorizontalDistributeCenter,
+  AlignStartHorizontal,
+  AlignStartVertical,
+  AlignVerticalDistributeCenter,
+  BringToFront,
+  ChevronDown,
+  ChevronUp,
+  ClipboardPaste,
+  Copy as CopyIcon,
+  ExternalLink,
+  FileBox,
+  Frame,
+  Image as ImageIcon,
+  ImageOff,
+  Link,
+  Lock as LockIcon,
+  LockOpen,
+  PanelBottom,
+  PanelLeft,
+  PanelRight,
+  PanelTop,
+  Pencil,
+  Scissors,
+  SendToBack,
+  Sparkles,
+  SquareDashed,
+  StickyNote,
+  Tag,
+  Trash2,
+} from 'lucide-react'
+import type { MutableRefObject } from 'react'
+import type { ResolvedTheme } from '../../hooks/useThemeMode.js'
+import { hasClipboardFragment } from '../../lib/clipboard-store.js'
+import type { AlignableBox, BoxMove } from './align.js'
+import { alignBoxes, distributeBoxes } from './align.js'
+import type { FileRefOption } from './CanvasPickerDialog.js'
+import { ContextMenu, type ContextMenuItem } from './ContextMenu.js'
+import type { EditorCommand } from './commands.js'
+import { type GestureState, reduceGesture } from './gestures.js'
+import type { Point } from './viewport.js'
+
+/** Open right-click menu: screen position (root-relative) + hit target. */
+export interface ContextMenuTarget {
+  readonly x: number
+  readonly y: number
+  readonly nodeId: string | undefined
+  readonly edgeId: string | undefined
+  readonly point: Point
+}
+
+/**
+ * `point` (canvas space) is present when creation came from the
+ * empty-space context menu: the user already chose WHERE, so the node
+ * lands there instead of the viewport-center free spot.
+ */
+export type LinkDialogState =
+  | { readonly mode: 'create'; readonly point?: Point }
+  | { readonly mode: 'edit'; readonly nodeId: string }
+
+export type CanvasPickerState =
+  | { readonly mode: 'create'; readonly point?: Point }
+  | { readonly mode: 'retarget'; readonly nodeId: string }
+
+export interface CanvasContextMenuProps {
+  readonly contextMenu: ContextMenuTarget
+  readonly setContextMenu: (target: ContextMenuTarget | null) => void
+  readonly canvas: SpatialCanvas
+  readonly canvasRef: MutableRefObject<SpatialCanvas>
+  readonly theme: ResolvedTheme
+  readonly gestureState: GestureState
+  readonly applyResult: (result: ReturnType<typeof reduceGesture>) => void
+  readonly isEdgeLocked: (edgeId: string) => boolean
+  readonly onToggleEdgeLock?: (edgeId: string, locked: boolean) => void
+  readonly edgeLockEnabled: boolean
+  readonly setEdgeLabelEditId: (id: string | null) => void
+  readonly setSelectedEdgeId: (id: string | null) => void
+  readonly setGroupLabelEditId: (id: string | null) => void
+  readonly pasteClipboard: (at?: Point) => boolean
+  readonly createNodeAt: (point: Point) => void
+  readonly createGroupAtViewportCenter: (at?: Point) => void
+  readonly setLinkDialog: (state: LinkDialogState | null) => void
+  readonly fileRefOptions?: readonly FileRefOption[]
+  readonly setCanvasPicker: (state: CanvasPickerState | null) => void
+  readonly onAddImage?: (file: File) => Promise<string | undefined>
+  readonly pendingImagePointRef: MutableRefObject<Point | null>
+  readonly imageInputRef: MutableRefObject<HTMLInputElement | null>
+  readonly pendingBackgroundGroupIdRef: MutableRefObject<string | null>
+  readonly isLocked: (nodeId: string) => boolean
+  readonly lockEnabled: boolean
+  readonly onToggleNodeLock?: (nodeId: string, locked: boolean) => void
+  readonly applyBoxMoves: (moves: readonly BoxMove[]) => boolean
+  readonly selectedAlignableBoxes: () => AlignableBox[]
+  readonly extraIds: ReadonlySet<string>
+  readonly selectedId: string | null
+  readonly groupSelection: (memberIds: readonly string[]) => void
+  readonly isImageFileRef?: (file: string) => boolean
+  readonly onOpenFileRef?: (file: string, subpath?: string) => void
+  readonly openLinkNode: (node: Extract<SpatialNode, { type: 'link' }>) => void
+  readonly copySelection: () => ClipboardFragment | null
+  readonly deleteSelectionAsBatch: () => boolean
+  readonly duplicateSelection: () => boolean
+  readonly reorderSelection: (placement: 'forward' | 'backward' | 'front' | 'back') => void
+}
+
+export function CanvasContextMenu({
+  contextMenu,
+  setContextMenu,
+  canvas,
+  canvasRef,
+  theme,
+  gestureState,
+  applyResult,
+  isEdgeLocked,
+  onToggleEdgeLock,
+  edgeLockEnabled,
+  setEdgeLabelEditId,
+  setSelectedEdgeId,
+  setGroupLabelEditId,
+  pasteClipboard,
+  createNodeAt,
+  createGroupAtViewportCenter,
+  setLinkDialog,
+  fileRefOptions,
+  setCanvasPicker,
+  onAddImage,
+  pendingImagePointRef,
+  imageInputRef,
+  pendingBackgroundGroupIdRef,
+  isLocked,
+  lockEnabled,
+  onToggleNodeLock,
+  applyBoxMoves,
+  selectedAlignableBoxes,
+  extraIds,
+  selectedId,
+  groupSelection,
+  isImageFileRef,
+  onOpenFileRef,
+  openLinkNode,
+  copySelection,
+  deleteSelectionAsBatch,
+  duplicateSelection,
+  reorderSelection,
+}: CanvasContextMenuProps) {
+  return (
+    <ContextMenu
+      x={contextMenu.x}
+      y={contextMenu.y}
+      onClose={() => setContextMenu(null)}
+      items={(() => {
+        const node =
+          contextMenu.nodeId === undefined
+            ? undefined
+            : canvas.nodes.find((n) => n.id === contextMenu.nodeId)
+        const edge =
+          contextMenu.edgeId === undefined
+            ? undefined
+            : canvas.edges.find((entry) => entry.id === contextMenu.edgeId)
+        // The swatch chips preview the CURRENT mode's preset strokes so
+        // the picker shows what will actually render; the stored value
+        // stays the semantic slot ('1'..'6'), never a resolved hex.
+        const presetSwatches = (theme === 'dark' ? SPATIAL_DARK_PALETTE : SPATIAL_LIGHT_PALETTE)
+          .presets
+        const presetEntries: readonly {
+          readonly key: SpatialPresetKey
+          readonly name: string
+        }[] = [
+          { key: '1', name: 'Red' },
+          { key: '2', name: 'Orange' },
+          { key: '3', name: 'Yellow' },
+          { key: '4', name: 'Green' },
+          { key: '5', name: 'Cyan' },
+          { key: '6', name: 'Purple' },
+        ]
+        const colorRow = (
+          current: CanvasColor | undefined,
+          apply: (color: CanvasColor | undefined) => void,
+        ) => ({
+          kind: 'options' as const,
+          label: 'Color',
+          options: [
+            {
+              label: 'default',
+              ariaLabel: 'Default',
+              icon: <SquareDashed />,
+              selected: current === undefined,
+              onSelect: () => apply(undefined),
+            },
+            ...presetEntries.map((entry) => ({
+              label: entry.key,
+              ariaLabel: entry.name,
+              icon: (
+                // Paint-critical props are inline, not utility classes:
+                // a default-inline span ignores width/height entirely
+                // (it laid out 0x0 live), and the chip must also paint
+                // where the app stylesheet is absent.
+                <span
+                  style={{
+                    display: 'block',
+                    width: 14,
+                    height: 14,
+                    borderRadius: '50%',
+                    backgroundColor: presetSwatches[entry.key].stroke,
+                  }}
+                />
+              ),
+              selected: current === entry.key,
+              onSelect: () => apply(entry.key),
+            })),
+          ],
+          // The JSON Canvas color union is presets OR a 6-digit hex;
+          // the native color input covers the hex half the swatches
+          // cannot.
+          customColor: {
+            value: current !== undefined && current.startsWith('#') ? current : '#808080',
+            ariaLabel: 'Custom color',
+            selected: current !== undefined && current.startsWith('#'),
+            onPick: (hex: string) => apply(hex as CanvasColor),
+          },
+        })
+        if (node === undefined && edge !== undefined) {
+          // A locked edge offers exactly one action. Everything else in
+          // this branch — delete, label, arrowheads, sides, colour —
+          // is a mutation the lock exists to refuse.
+          if (isEdgeLocked(edge.id)) {
+            return [
+              {
+                label: 'Unlock',
+                icon: <LockOpen />,
+                onSelect: () => onToggleEdgeLock?.(edge.id, false),
+              },
+            ]
+          }
+          // Property pickers are inline option rows (one tap per
+          // choice, menu stays open) — a cycling item costs an
+          // open-tap-reopen per step. Sections group the menu:
+          // actions, then properties, then the destructive entry.
+          // Arrow direction reads the JSON Canvas defaults (fromEnd
+          // none, toEnd arrow).
+          const fromEnd = edge.fromEnd ?? 'none'
+          const toEnd = edge.toEnd ?? 'arrow'
+          const arrowStates = [
+            { label: '→', ariaLabel: 'Forward', fromEnd: 'none', toEnd: 'arrow' },
+            { label: '↔', ariaLabel: 'Both', fromEnd: 'arrow', toEnd: 'arrow' },
+            { label: '←', ariaLabel: 'Backward', fromEnd: 'arrow', toEnd: 'none' },
+            { label: '−', ariaLabel: 'None', fromEnd: 'none', toEnd: 'none' },
+          ] as const
+          const applyEdgeCommand = (command: EditorCommand) =>
+            applyResult({ state: { kind: 'idle' }, commands: [command] })
+          // Excel-border-style side pickers: a rectangle with the
+          // pinned side emphasized; dashed = unpinned (auto).
+          const SIDES = [
+            { label: 'auto', ariaLabel: 'Auto', icon: <SquareDashed />, side: undefined },
+            { label: 'top', ariaLabel: 'Top', icon: <PanelTop />, side: 'top' },
+            { label: 'right', ariaLabel: 'Right', icon: <PanelRight />, side: 'right' },
+            { label: 'bottom', ariaLabel: 'Bottom', icon: <PanelBottom />, side: 'bottom' },
+            { label: 'left', ariaLabel: 'Left', icon: <PanelLeft />, side: 'left' },
+          ] as const
+          const sideRow = (endpoint: 'from' | 'to') => {
+            const current = endpoint === 'from' ? edge.fromSide : edge.toSide
+            return {
+              kind: 'options' as const,
+              label: endpoint === 'from' ? 'From side' : 'To side',
+              options: SIDES.map((entry) => ({
+                label: entry.label,
+                icon: entry.icon,
+                ariaLabel: entry.ariaLabel,
+                selected: current === entry.side,
+                onSelect: () =>
+                  applyEdgeCommand({
+                    kind: 'set-edge-side',
+                    id: edge.id,
+                    endpoint,
+                    side: entry.side,
+                  }),
+              })),
+            }
+          }
+          return [
+            {
+              label: 'Edit label',
+              icon: <Tag />,
+              onSelect: () => setEdgeLabelEditId(edge.id),
+            },
+            { kind: 'separator' as const },
+            {
+              kind: 'options' as const,
+              label: 'Arrows',
+              options: arrowStates.map((state) => ({
+                label: state.label,
+                ariaLabel: state.ariaLabel,
+                selected: state.fromEnd === fromEnd && state.toEnd === toEnd,
+                onSelect: () =>
+                  applyEdgeCommand({
+                    kind: 'set-edge-ends',
+                    id: edge.id,
+                    fromEnd: state.fromEnd,
+                    toEnd: state.toEnd,
+                  }),
+              })),
+            },
+            sideRow('from'),
+            sideRow('to'),
+            colorRow(edge.color, (color) =>
+              applyEdgeCommand({ kind: 'set-edge-color', id: edge.id, color }),
+            ),
+            ...(edgeLockEnabled
+              ? [
+                  {
+                    label: 'Lock',
+                    icon: <LockIcon />,
+                    onSelect: () => {
+                      onToggleEdgeLock?.(edge.id, true)
+                      setSelectedEdgeId(null)
+                    },
+                  },
+                ]
+              : []),
+            { kind: 'separator' as const },
+            {
+              label: 'Delete',
+              icon: <Trash2 />,
+              danger: true,
+              onSelect: () => {
+                applyResult({
+                  state: { kind: 'idle' },
+                  commands: [{ kind: 'delete-edge', id: edge.id } as const],
+                })
+                setSelectedEdgeId(null)
+              },
+            },
+          ]
+        }
+        if (node === undefined) {
+          // The same creation set as the dock's + menu, anchored at
+          // the click point — "here" is exactly the information the
+          // bottom dock cannot express.
+          const emptyItems: ContextMenuItem[] = [
+            ...(hasClipboardFragment()
+              ? [
+                  {
+                    label: 'Paste here',
+                    icon: <ClipboardPaste />,
+                    onSelect: () => {
+                      pasteClipboard(contextMenu.point)
+                    },
+                  },
+                  { kind: 'separator' } as const,
+                ]
+              : []),
+            {
+              label: 'Add note here',
+              icon: <StickyNote />,
+              onSelect: () => createNodeAt(contextMenu.point),
+            },
+            {
+              label: 'Add link here',
+              icon: <Link />,
+              onSelect: () => setLinkDialog({ mode: 'create', point: contextMenu.point }),
+            },
+            {
+              label: 'Add group here',
+              icon: <Frame />,
+              onSelect: () => createGroupAtViewportCenter(contextMenu.point),
+            },
+          ]
+          if (fileRefOptions !== undefined) {
+            emptyItems.push({
+              label: 'Add canvas here',
+              icon: <FileBox />,
+              onSelect: () => setCanvasPicker({ mode: 'create', point: contextMenu.point }),
+            })
+          }
+          if (onAddImage !== undefined) {
+            emptyItems.push({
+              label: 'Add image here',
+              icon: <ImageIcon />,
+              onSelect: () => {
+                pendingImagePointRef.current = contextMenu.point
+                imageInputRef.current?.click()
+              },
+            })
+          }
+          // Tidy needs a second node to tidy AGAINST — the item appears
+          // only once it can do something, like Align/Distribute above.
+          if (canvas.nodes.length >= 2) {
+            emptyItems.push({ kind: 'separator' })
+            emptyItems.push({
+              label: 'Tidy canvas',
+              icon: <Sparkles />,
+              onSelect: () =>
+                applyBoxMoves(tidyNodes(canvasRef.current.nodes, { locked: isLocked })),
+            })
+          }
+          return emptyItems
+        }
+        const items: ContextMenuItem[] = []
+        if (node.type === 'group') {
+          items.push({
+            label: 'Edit label',
+            icon: <Tag />,
+            onSelect: () => setGroupLabelEditId(node.id),
+          })
+          if (onAddImage !== undefined) {
+            items.push({
+              label: 'Set background image',
+              icon: <ImageIcon />,
+              onSelect: () => {
+                pendingBackgroundGroupIdRef.current = node.id
+                imageInputRef.current?.click()
+              },
+            })
+          }
+          if (node.background !== undefined) {
+            const background = node.background
+            const applyStyle = (backgroundStyle: 'cover' | 'ratio') =>
+              applyResult({
+                state: { kind: 'idle' },
+                commands: [
+                  { kind: 'set-group-background', id: node.id, background, backgroundStyle },
+                ],
+              })
+            items.push({
+              kind: 'options',
+              label: 'Background',
+              options: [
+                {
+                  label: 'Cover',
+                  ariaLabel: 'Cover',
+                  selected: node.backgroundStyle !== 'ratio',
+                  onSelect: () => applyStyle('cover'),
+                },
+                {
+                  label: 'Fit',
+                  ariaLabel: 'Fit',
+                  selected: node.backgroundStyle === 'ratio',
+                  onSelect: () => applyStyle('ratio'),
+                },
+              ],
+            })
+            items.push({
+              label: 'Remove background',
+              icon: <ImageOff />,
+              onSelect: () =>
+                applyResult({
+                  state: { kind: 'idle' },
+                  commands: [{ kind: 'set-group-background', id: node.id }],
+                }),
+            })
+          }
+          items.push({ kind: 'separator' })
+        }
+        // Framing an existing multi-selection is reached from any of
+        // its members — the frame encloses every selected node,
+        // including group frames: nesting is geometric in JSON Canvas,
+        // and containment moves already handle nested frames.
+        if (extraIds.size > 0) {
+          items.push({
+            label: 'Group selection',
+            icon: <Frame />,
+            onSelect: () => groupSelection([node.id, ...extraIds]),
+          })
+          items.push({ kind: 'separator' })
+        }
+        if (node.type === 'file' && isImageFileRef?.(node.file) !== true) {
+          if (onOpenFileRef !== undefined) {
+            items.push({
+              label: 'Open canvas',
+              icon: <ExternalLink />,
+              onSelect: () => onOpenFileRef(node.file, node.subpath),
+            })
+          }
+          if (fileRefOptions !== undefined) {
+            items.push({
+              label: 'Change target',
+              icon: <FileBox />,
+              onSelect: () => setCanvasPicker({ mode: 'retarget', nodeId: node.id }),
+            })
+          }
+          if (onOpenFileRef !== undefined || fileRefOptions !== undefined) {
+            items.push({ kind: 'separator' })
+          }
+        }
+        if (node.type === 'link') {
+          items.push({
+            label: 'Open link',
+            icon: <ExternalLink />,
+            onSelect: () => openLinkNode(node),
+          })
+          items.push({
+            label: 'Edit URL',
+            icon: <Pencil />,
+            onSelect: () => setLinkDialog({ mode: 'edit', nodeId: node.id }),
+          })
+          items.push({ kind: 'separator' })
+        }
+        if (node.type === 'text') {
+          items.push({
+            label: 'Edit text',
+            icon: <Pencil />,
+            onSelect: () => {
+              applyResult(
+                reduceGesture(gestureState, canvas, {
+                  type: 'start-text-edit',
+                  nodeId: node.id,
+                  text: node.text,
+                }),
+              )
+            },
+          })
+          // Same grouping rule as the edge menu: the destructive
+          // entry sits in its own section.
+          items.push({ kind: 'separator' })
+        }
+        // Touch path to Cmd/Ctrl+D (see shortcuts.ts). The menu's
+        // right-click already made this node the primary selection,
+        // so the shared handler clones the full multi-selection.
+        // A locked node's menu offers exactly one action: unlock.
+        // Showing Delete/Edit next to a lock the user deliberately
+        // set would make the lock read as decorative.
+        if (isLocked(node.id)) {
+          return [
+            {
+              label: 'Unlock',
+              icon: <LockOpen />,
+              onSelect: () => onToggleNodeLock?.(node.id, false),
+            },
+          ]
+        }
+        items.push({
+          label: 'Copy',
+          icon: <CopyIcon />,
+          onSelect: () => {
+            copySelection()
+          },
+        })
+        items.push({
+          label: 'Cut',
+          icon: <Scissors />,
+          onSelect: () => {
+            // Menu path mirrors the native cut: copy, then remove.
+            if (copySelection() !== null) deleteSelectionAsBatch()
+          },
+        })
+        items.push({
+          label: 'Duplicate',
+          icon: <CopyIcon />,
+          onSelect: () => {
+            duplicateSelection()
+          },
+        })
+        items.push(
+          colorRow(node.color, (color) => {
+            // Recoloring FROM a multi-selection styles the whole
+            // selected AREA: every member, and every edge that runs
+            // between two members — the closest executable reading of
+            // "select a region, recolor it". An edge leaving the
+            // selection keeps its color (only one endpoint is inside),
+            // and a target outside the selection styles itself alone.
+            const members = new Set(selectedId !== null ? [selectedId, ...extraIds] : [])
+            const nodeTargets = members.has(node.id) ? [...members] : [node.id]
+            const edgeTargets =
+              nodeTargets.length > 1
+                ? canvas.edges.filter(
+                    (edge) =>
+                      !isEdgeLocked(edge.id) &&
+                      members.has(edge.fromNode) &&
+                      members.has(edge.toNode),
+                  )
+                : []
+            applyResult({
+              state: { kind: 'idle' },
+              commands: [
+                ...nodeTargets.map((id) => ({ kind: 'set-node-color' as const, id, color })),
+                ...edgeTargets.map((edge) => ({
+                  kind: 'set-edge-color' as const,
+                  id: edge.id,
+                  color,
+                })),
+              ],
+            })
+          }),
+        )
+        // Z-order as one-tap options — the touch path to the [ / ]
+        // keyboard shortcuts (see shortcuts.ts). Not a picker: no
+        // option is ever "selected", each tap applies a move.
+        items.push({
+          kind: 'options',
+          label: 'Order',
+          options: [
+            {
+              label: 'back',
+              ariaLabel: 'Send to back',
+              icon: <SendToBack />,
+              selected: false,
+              onSelect: () => reorderSelection('back'),
+            },
+            {
+              label: 'backward',
+              ariaLabel: 'Send backward',
+              icon: <ChevronDown />,
+              selected: false,
+              onSelect: () => reorderSelection('backward'),
+            },
+            {
+              label: 'forward',
+              ariaLabel: 'Bring forward',
+              icon: <ChevronUp />,
+              selected: false,
+              onSelect: () => reorderSelection('forward'),
+            },
+            {
+              label: 'front',
+              ariaLabel: 'Bring to front',
+              icon: <BringToFront />,
+              selected: false,
+              onSelect: () => reorderSelection('front'),
+            },
+          ],
+        })
+        // Align needs a second box to align TO, and distribute needs a
+        // middle one to place — so each row appears only once its
+        // action means something, rather than sitting there inert.
+        const alignableCount = extraIds.size + 1
+        if (alignableCount >= 2) {
+          items.push({
+            kind: 'options',
+            label: 'Align',
+            options: (
+              [
+                ['left', 'Align left', <AlignStartVertical key="l" />],
+                ['center-x', 'Align centre horizontally', <AlignCenterVertical key="cx" />],
+                ['right', 'Align right', <AlignEndVertical key="r" />],
+                ['top', 'Align top', <AlignStartHorizontal key="t" />],
+                ['center-y', 'Align centre vertically', <AlignCenterHorizontal key="cy" />],
+                ['bottom', 'Align bottom', <AlignEndHorizontal key="b" />],
+              ] as const
+            ).map(([mode, ariaLabel, icon]) => ({
+              label: mode,
+              ariaLabel,
+              icon,
+              selected: false,
+              onSelect: () => applyBoxMoves(alignBoxes(selectedAlignableBoxes(), mode)),
+            })),
+          })
+        }
+        if (alignableCount >= 3) {
+          items.push({
+            kind: 'options',
+            label: 'Distribute',
+            options: (
+              [
+                [
+                  'horizontal',
+                  'Distribute horizontally',
+                  <AlignHorizontalDistributeCenter key="h" />,
+                ],
+                ['vertical', 'Distribute vertically', <AlignVerticalDistributeCenter key="v" />],
+              ] as const
+            ).map(([axis, ariaLabel, icon]) => ({
+              label: axis,
+              ariaLabel,
+              icon,
+              selected: false,
+              onSelect: () => applyBoxMoves(distributeBoxes(selectedAlignableBoxes(), axis)),
+            })),
+          })
+        }
+        if (alignableCount >= 2) {
+          items.push({
+            label: 'Tidy',
+            icon: <Sparkles />,
+            onSelect: () =>
+              applyBoxMoves(
+                tidyNodes(canvasRef.current.nodes, {
+                  scope: new Set([node.id, ...extraIds]),
+                  locked: isLocked,
+                }),
+              ),
+          })
+        }
+        if (lockEnabled) {
+          items.push({
+            label: 'Lock',
+            icon: <LockIcon />,
+            onSelect: () => onToggleNodeLock?.(node.id, true),
+          })
+        }
+        items.push({ kind: 'separator' })
+        items.push({
+          label: 'Delete',
+          icon: <Trash2 />,
+          danger: true,
+          onSelect: () => {
+            applyResult(
+              reduceGesture(gestureState, canvas, {
+                type: 'delete-selection',
+                nodeId: node.id,
+              }),
+            )
+          },
+        })
+        return items
+      })()}
+    />
+  )
+}
